Added support for arrays of custom fields
The previous implementation would take the first instance of a custom field and return that, which is definitely the most common case... but not the only one. I extended the custom field syntax to allow an options object. You can now pass a `keepArray` option to return an array of nodes (a la the default `categories` behavior).